Ajeitando inconsistencias

+++ app/Controllers/Http/ClientesController.js
@@ -1,7 +1,5 @@
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
-const Empregados = use('App/Models/Empregados');
-/** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
-const Clientes = use('App/Models/Clientes');
+const User = use('App/Models/User');
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
 const Endereco = use('App/Models/Endereco');
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
@@ -9,9 +7,9 @@ const Telefone = use('App/Models/Telefone');
 
 class ClientesController {
   async store({ request, response }) {
-    const cliente = await Clientes.create(
-      request.only(['nome', 'email', 'cpf'])
-    );
+    const cliente = await User.create(request.only(['nome', 'email', 'cpf']), {
+      tipo: 'cliente',
+    });
     const endereco = await Endereco.create(
       request.only([
         'cep',
@@ -34,14 +32,8 @@ class ClientesController {
       .json({ newUser: { cliente, endereco, telefone } });
   }
 
-  async index({ auth, response }) {
-    const empregadoBoss = await Empregados.findOrFail('user_id', auth.user_id);
-
-    if (empregadoBoss.tipo !== 'ADM') {
-      return response.status(401).json('Você não tem autorização');
-    }
-
-    const cliente = await Clientes.all();
+  async index() {
+    const cliente = await User.findByOrFail('tipo', 'cliente');
 
     if (!cliente) {
       return 'Não há clientes cadastrados';
@@ -52,8 +44,22 @@ class ClientesController {
     return cliente;
   }
 
+  async show({ params, response }) {
+    const cliente = await User.findOrFail(params.id);
+
+    if (cliente.tipo !== 'cliente') {
+      return response.status(400).json('Esse user não é um cliente');
+    }
+
+    return cliente;
+  }
+
   async update({ params, request, response }) {
-    const cliente = await Clientes.findOrFail(params.id);
+    const cliente = await User.findOrFail(params.id);
+
+    if (cliente.tipo !== 'cliente') {
+      return response.status(400).json('Esse user não é um cliente');
+    }
 
     cliente.merge(request.all());
 
@@ -63,13 +69,17 @@ class ClientesController {
   }
 
   async destroy({ params, response, auth }) {
-    const empregadoBoss = await Empregados.findOrFail('user_id', auth.user_id);
+    const empregadoBoss = await User.findOrFail('user_id', auth.user_id);
 
     if (empregadoBoss.tipo !== 'ADM') {
       return response.status(401).json('Você não tem autorização');
     }
 
-    const cliente = await Clientes.findOrFail(params.id);
+    const cliente = await User.findOrFail(params.id);
+
+    if (cliente.tipo !== 'cliente') {
+      return response.status(400).json('Esse user não é um cliente');
+    }
 
     await cliente.delete();
