+++ app/Controllers/Http/EmpregadoController.js
@@ -1,5 +1,5 @@
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
-const Empregados = use('App/Models/Empregados');
+const User = use('App/Models/User');
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
 const Endereco = use('App/Models/Endereco');
 /** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
@@ -7,7 +7,7 @@ const Telefone = use('App/Models/Telefone');
 
 class EmpregadoController {
   async store({ request, auth, response }) {
-    const empregadoBoss = await Empregados.findByOrFail(auth.user.id);
+    const empregadoBoss = await User.findByOrFail(auth.user.id);
 
     if (empregadoBoss.tipo !== 'ADM') {
       return response
@@ -15,8 +15,10 @@ class EmpregadoController {
         .json('Você não tem autorização para cadastrar funcionários');
     }
 
-    const empregados = await Empregados.create(
-      request.only(['nome', 'email', 'cpf', 'password'])
+    const empregados = await User.create(
+      request.only(['nome', 'email', 'cpf', 'password'], {
+        tipo: 'empregado',
+      })
     );
 
     const endereco = await Endereco.create(
@@ -42,15 +44,15 @@ class EmpregadoController {
   }
 
   async index({ auth, response }) {
-    const empregadoBoss = await Empregados.findOrFail('user_id', auth.user_id);
+    const empregadoBoss = await User.findOrFail('user_id', auth.user_id);
 
     if (empregadoBoss.tipo !== 'ADM') {
       return response
         .status(401)
-        .json('Você não tem autorização para cadastrar funcionários');
+        .json('Você não tem autorização para lista funcionários');
     }
 
-    const empregados = await Empregados.all();
+    const empregados = await User.findOrFail('tipo', 'empregado');
 
     if (!empregados) {
       return 'Não há Empregados cadastrados';
@@ -62,23 +64,31 @@ class EmpregadoController {
   }
 
   async update({ params, request, response }) {
-    const enderecos = await Empregados.findOrFail(params.id);
+    const empregado = await User.findOrFail(params.id);
 
-    enderecos.merge(request.all());
+    if (empregado.tipo !== 'empregado') {
+      return response.status(400).json('Esse user não é um empregado');
+    }
 
-    await enderecos.save();
+    empregado.merge(request.all());
 
-    return response.status(201).json(enderecos);
+    await empregado.save();
+
+    return response.status(201).json(empregado);
   }
 
   async destroy({ params, response, auth }) {
-    const empregadoBoss = await Empregados.findOrFail('user_id', auth.user_id);
+    const empregadoBoss = await User.findOrFail('user_id', auth.user_id);
 
     if (empregadoBoss.tipo !== 'ADM') {
       return response.status(401).json('Você não tem autorizado');
     }
 
-    const empregado = await Empregados.findOrFail(params.id);
+    const empregado = await User.findOrFail(params.id);
+
+    if (empregado.tipo !== 'empregado') {
+      return response.status(400).json('Esse user não é um emprego');
+    }
 
     await empregado.delete();
